marking

+++ lib/Models/score_tile.dart
@@ -3,7 +3,7 @@ import 'package:quiz_app/Constants/images.dart';
 class ScoreTile {
   String leadingImg;
   String title;
-  double mark;
+  int mark;
   ScoreTile({
     this.leadingImg = testimg,
     required this.title,
